% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllUtterancesAndObjects.R
\docType{data}
\name{allFeatureTypesNew}
\alias{allFeatureTypesNew}
\title{All feature types}
\format{
An object of class \code{character} of length 3.
}
\usage{
allFeatureTypesNew
}
\description{
A vector of all object features: ('shape','pattern','color')
}
\keyword{datasets}
