% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllUtterancesAndObjects.R
\docType{data}
\name{allObjects}
\alias{allObjects}
\title{All objects}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 27 rows and 3 columns.
}
\usage{
allObjects
}
\description{
The matrix contains \code{3^3} types of objects.

It essentially specifies the 3 feature expressions for each object
thus, the matrix maps objects to matching utterances.
The matrix contains strings of all possible feature values of the objects.

The index mapping of the strings are contained in the matrix
\code{\link{allObjectsToUtterancesMappings}}.
}
\examples{
\dontrun{
output:
     shape   pattern  color
[1,] cloud    solid   blue
[2,] circle   solid   blue
[3,] square   solid   blue
[4,] cloud   striped  blue
[5,] circle  striped  blue
[6,] square  striped  blue
[7,] cloud   dotted   blue
[8,] circle  dotted   blue
[9,] square  dotted   blue
[10,] cloud    solid    red
[11,] circle   solid    red
[12,] square   solid    red
[13,] cloud   striped   red
[14,] circle  striped   red
[15,] square  striped   red
[16,] cloud   dotted    red
[17,] circle  dotted    red
[18,] square  dotted    red
[19,] cloud    solid    green
[20,] circle   solid    green
[21,] square   solid    green
[22,] cloud   striped   green
[23,] circle  striped   green
[24,] square  striped   green
[25,] cloud   dotted    green
[26,] circle  dotted    green
[27,] square  dotted    green
}
}
\keyword{datasets}
