% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllUtterancesAndObjects.R
\docType{data}
\name{allObjectsToUtterancesMappings}
\alias{allObjectsToUtterancesMappings}
\title{All objects to utterances mappings}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 27 rows and 3 columns.
}
\usage{
allObjectsToUtterancesMappings
}
\description{
The matrix contains \code{3^3} types of objects.

It essentially specifies the 3 feature expressions for each object
thus, the matrix maps objects to matching utterances.
The matrix contains the index mapping of all possible feature values:

cloud = 1, circle = 2, square = 3

solid = 4, striped = 5, dotted = 6

blue = 7, red = 8, green = 9

The strings of the index mapping are contained in the matrix
\code{\link{allObjects}}.
}
\examples{
\dontrun{
output:

      [,1] [,2] [,3]
[1,]    1    4    7
[2,]    2    4    7
[3,]    3    4    7
[4,]    1    5    7
[5,]    2    5    7
[6,]    3    5    7
[7,]    1    6    7
[8,]    2    6    7
[9,]    3    6    7
[10,]    1    4    8
[11,]    2    4    8
[12,]    3    4    8
[13,]    1    5    8
[14,]    2    5    8
[15,]    3    5    8
[16,]    1    6    8
[17,]    2    6    8
[18,]    3    6    8
[19,]    1    4    9
[20,]    2    4    9
[21,]    3    4    9
[22,]    1    5    9
[23,]    2    5    9
[24,]    3    5    9
[25,]    1    6    9
[26,]    2    6    9
[27,]    3    6    9
}

}
\keyword{datasets}
