% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllUtterancesAndObjects.R
\docType{data}
\name{allUtterances}
\alias{allUtterances}
\title{All utterances}
\format{
An object of class \code{character} of length 9.
}
\usage{
allUtterances
}
\description{
A vector that contains all feature values:

('cloud', 'circle', 'square', 'solid', 'striped', 'dotted', 'blue', 'red', 'green')
}
\details{
The difference between this matrix and the \code{\link{allUtterancesNew}} is that
the 'dotted' pattern is called \emph{polka-dotted}.
}
\keyword{datasets}
