% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllUtterancesAndObjects.R
\docType{data}
\name{allUtterancesNew}
\alias{allUtterancesNew}
\title{All utterances (new)}
\format{
An object of class \code{character} of length 9.
}
\usage{
allUtterancesNew
}
\description{
A vector that contains all feature values:

('cloud', 'circle', 'square', 'solid', 'striped', 'polka-dotted', 'blue', 'red', 'green')
}
\details{
The difference between this matrix and the \code{\link{allUtterances}} is that
the 'polka-dotted' pattern is called \emph{dotted}.
}
\keyword{datasets}
