% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSA_StratUttOptimization.R
\name{determineSpeakerPostListPrefs}
\alias{determineSpeakerPostListPrefs}
\title{Determine speaker's inference of the posterior listener preferences}
\usage{
determineSpeakerPostListPrefs(
  currentObjects,
  featureUtt,
  softPrefValue,
  notObeyInst,
  alpha
)
}
\arguments{
\item{currentObjects}{A vector of three values in \code{{1,...,27}} specifying the target and the other two objects in the scene.

The target is the first object in the vector \code{(index = 1)}.}

\item{featureUtt}{One of the values \code{{1,2,3}} specifying which feature is uttered (i.e. shape = 1 / texture = 2 / or color = 3).}

\item{softPrefValue}{A parameter value between \code{[0,infinity)} (The larger the higher the tendency towards uniform liking).

Value reflects how categorical the listener's preferences are:

\strong{0:} The listener always picks her preferred object.

If the listener prefers \emph{red} objects, she will always pick the \emph{red} object in the scene.

\strong{infinity:} It is as likely for the listener to pick \emph{green}, \emph{blue} or \emph{red} objects.}

\item{notObeyInst}{Determines the extent to which the instruction of the speaker is obeyed by the listener.

(0 = full obedience, infinity = full instruction ignorance).

\strong{Example:}

\strong{0:} Listener always picks \emph{red} objects following the utterance \emph{"red"}.

\strong{infinity:} Listener as likely to pick \emph{green, blue} or \emph{red} objects even if the utterance is \emph{"red"}.}

\item{alpha}{A parameter value between 0 and 1.

Exponential scaling of the speaker choosing the utterance that maximizes the chance of the listener getting the target object right.}
}
\value{
A vector of length 9. It contains the speaker's inference of the feature value preferences of the listener.
}
\description{
Full-RSA

This function calculates the speaker's posterior guess of the feature value preferences of the listener.
That means how the speaker infers the preferences of the listener based on the object choice.
}
