% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllUtterancesAndObjects.R
\name{determineValidUtterances}
\alias{determineValidUtterances}
\title{Determine valid utterances}
\usage{
determineValidUtterances(currentObjects)
}
\arguments{
\item{currentObjects}{A vector of three values in range \code{{1,...,27}} specifying the target and the other two objects.

The target is the first object in the vector \code{(index = 1)}.}
}
\value{
A vector of utterances that correspond to all feature values present in the current objects in the scene.
}
\description{
The relevant utterances are determined based on the features of the current objects present in the scene.
}
\examples{
\donttest{currentObjects <- c(1,2,3)}

\donttest{determineValidUtterances(currentObjects)}

output: [1] 1 2 3 4 7

To see which objects are present in the scene run:
\donttest{allObjects[currentObjects,]}

output:

      shape    pattern color
[1,] "cloud"  "solid" "blue"
[2,] "circle" "solid" "blue"
[3,] "square" "solid" "blue"

Based on this output you can see that the indices in the validUtterances vector refer to:
cloud = 1, circle = 2, square = 3, solid = 4 and blue = 7.

}
