% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConstCodeStratUtt.R
\name{getConstellationCode}
\alias{getConstellationCode}
\title{Get object constellation code}
\usage{
getConstellationCode(objectConstellation, chosenFeature)
}
\arguments{
\item{objectConstellation}{A vector of three values in range \code{{1,...,27}} specifying the target and the other two objects in the scene.

The target is the first object in the vector \code{(index = 1)}.}

\item{chosenFeature}{A value \code{{1,2 or 3}} that specifies, which one of the three feature values (shape, texture or color)
of the chosen object was used as the utterance.

\strong{Example:}

utterance was \emph{"red"}, means chosen feature is \emph{color} (\code{chosenFeature = 3}).}
}
\value{
A list of three rows (length of the objectConstellation vector). Each row..
}
\description{
This function determines the ambiguity classes.
These are determined by answering this question: What properties does target object share with other objects?

This allows you to group together the trials that have the same levels of ambiguity.
}
\examples{
getConstellationCode(objectConstellation, chosenFeature)

output:
[[1]]
[1] 1 1 3 2 3 2

[[2]]
[1] 1 2 3

[[3]]
[1] 1 2 3

To see which object are in the object constellation run:

\donttest{allObjects[objecConstellation,]}

output:
     shape   pattern  color
[1,] "cloud"  "solid" "blue"
[2,] "circle" "solid" "blue"
[3,] "square" "solid" "blue"
}
