% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRSA_StratUtt.R
\name{getPreferencesPrior}
\alias{getPreferencesPrior}
\title{Get prior preferences of the listener}
\usage{
getPreferencesPrior(targetFeature)
}
\arguments{
\item{targetFeature}{A value between 1 and 3, specifying which feature type- color, shape, or pattern- is considered (for preferences).}
}
\value{
A vector of length 9. It contains a uniform prior over the three features of the specified feature type and zeros for the other feature values.
}
\description{
Simple RSA
}
\examples{
\donttest{getPreferencesPrior(targetFeature)}

output:
[1] 0.33 0.33 0.33 0.00 0.00 0.00 0.00 0.00 0.00

 @details
 This function is used in X9.
}
