% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSA_StratUtt.R
\name{getSpeakerUtterancePriors}
\alias{getSpeakerUtterancePriors}
\title{Get the speaker's utterances' priors}
\usage{
getSpeakerUtterancePriors(validUtterances)
}
\arguments{
\item{validUtterances}{A vector of utterances that correspond to all feature values present
in the current objects in the scene.

For example, it only makes sense to utter \emph{"red"} in a scene if there are \emph{red} objects present.}
}
\value{
A vector of priors as numeric values.
}
\description{
Full-RSA

This function determines the prior utterance preferences of the speaker.
}
\examples{
\donttest{getSpeakerUtterancePriors(validUtterances)}

output:
[1] 0.2 0.2 0.2 0.2 0.2
}
