% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRSA_StratUtt.R
\name{getSpeakerUtteranceUniformPrior}
\alias{getSpeakerUtteranceUniformPrior}
\title{Speaker's uniform priors for utterances}
\usage{
getSpeakerUtteranceUniformPrior(validUtterances)
}
\arguments{
\item{validUtterances}{A vector of utterances that correspond to all feature values present
 in the current objects in the scene.

For example, it only makes sense to utter \emph{"red"} in a scene if there are \emph{red} objects present.}
}
\value{
A vector of the same length as the validUtterances vector.

It contains numeric values of the prior utterance preferences of the speaker.
}
\description{
Simple RSA

Determines the prior utterance preferences of the speaker.
}
\examples{
\donttest{getSpeakerUtteranceUniformPrior(validUtterances)}

output:
 [1] 0.17 0.17 0.17 0.17 0.17 0.17
 
 @details
 This function is used in X9.
}
