% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConstCodeStratUtt.R
\name{getUtteranceChoiceConstellationCode}
\alias{getUtteranceChoiceConstellationCode}
\title{Get utterance-choice's constellation code}
\usage{
getUtteranceChoiceConstellationCode(objectConstellation)
}
\arguments{
\item{objectConstellation}{A vector of three values in range \code{{1,...,27}} specifying the target and the other two objects in the scene.

The target is the first object in the vector \code{(index = 1)}.}
}
\value{
A list of four rows.

\strong{First row:} resultCode

\strong{Second row:} featureOrder

\strong{Thrid row:} objectOrder

\strong{Fourth row:} featureValueOrder
}
\description{
Method that determines the utterance choice constellation code.

It also determines the resulting feature-order, the object-order and the feature-value-order.

The feature-order depends on the individual feature value ambiguities.
}
\examples{
\donttest{getUtteranceChoiceConstellationCode(objectConstellation)}

output:
[[1]]
[1] "331"

[[2]]
[1] 3 2 1

[[3]]
[1] 1 2 3

[[4]]
[1] 7 4 1 2 3

To see which object are in the object constellation run:
allObjects[objectConstellation,]

output:
     shape   pattern  color
[1,] "cloud"  "solid" "blue"
[2,] "circle" "solid" "blue"
[3,] "square" "solid" "blue"
}
