% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSA_StratUtt.R
\name{KLdivergence}
\alias{KLdivergence}
\title{Calculate Kullback-Leibler divergence}
\usage{
KLdivergence(p, q)
}
\arguments{
\item{p}{Probability distribution.}

\item{q}{Probability distribution.}
}
\value{
A scalar value (The divergence between p and q).
}
\description{
Divergence measure.

Simple KL divergence function- with small offset to tolerate \code{ p = 0 / q = 0}
}
\details{
Full-RSA and Simple-RSA
}
