% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSA_StratUtt.R
\name{literalListener}
\alias{literalListener}
\title{Literal listener function}
\usage{
literalListener(utterance, listenerObjectPreferences, uttToObjProbs)
}
\arguments{
\item{utterance}{The uttered word by the speaker that the listener hears.

An index referring to one of the values in the vector validUtterances.}

\item{listenerObjectPreferences}{One of the rows of the list of preference priors
for all valid utterances based on the object in the scene.

The list has as many rows as the length of the validUtterances vector + 1.

Each row in the list contains a vector of length 3, as there are three objects in the scene.

The extra row is for the case of no feature preferences whatsoever, i.e. uniform prior over all three objects in the scene.}

\item{uttToObjProbs}{A matrix. The rows map each possible utterance that corresponds to each present feature value
of the current objects. The columns represent the three objects in the scene.

This reflects the obedience-parameter and which objects match the respective utterance.
The matrix shows the probability that a certain object is chosen following a certain utterance, that is valid in the scene.
The number of rows of the matrix match the length of the validUtterances vector.}
}
\value{
P(obj | utt, listener's object preferences)

A Vector of length 3.
It includes the normalized probability of choosing each of the three objects in the scene,
given the utterance and the listener's object preferences.
}
\description{
Full-RSA

Literal listener function according to assigned listener's object preferences.
}
\examples{
listenerObjectPreferences <- objectPreferenceSoftPriors[[3]]
\donttest{literalListener(utterance, listenerObjectPreferences, uttToObjProbs)}

output:
[1] 1 0 0
}
