% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllUtterancesAndObjects.R
\name{mapObjectToUtterances}
\alias{mapObjectToUtterances}
\title{Map objects to utterances}
\usage{
mapObjectToUtterances(currentObjects)
}
\arguments{
\item{currentObjects}{A vector of three values in range \code{{1,...,27}} specifying the target and the other two objects.

The target is the first object in the vector \code{(index = 1)}.}
}
\value{
A 3x3 matrix. Values in the matrix are in range \code{{1-9}}.

\strong{rows:} Objects; \strong{columns:} Feature values of the objects
}
\description{
Determining the specific mapping of objects to utterances that applies given the present objects in the scene.
}
\details{
This function is only used in X9
}
\examples{
\donttest{currentObjects <- c(1,2,3)}
\donttest{mapObjectToUtterances(currentObjects)}

output:
      [,1] [,2] [,3]
[1,]    1    4    7
[2,]    2    4    7
[3,]    3    4    7

To see which objects are present in the scene run:
\donttest{allObjects[currentObjects,]}

output:

      shape   pattern  color
[1,] "cloud"  "solid" "blue"
[2,] "circle" "solid" "blue"
[3,] "square" "solid" "blue"

Based on this output you can see that the indices output matrix vector refer
to the following feature values:
cloud = 1, circle = 2, square = 3, solid = 4 and blue = 7.

}
