% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSA_StratUttOptimization.R
\name{RSAModelKLDiv3paramsOnlyAvailableFeatureValuesConsidered}
\alias{RSAModelKLDiv3paramsOnlyAvailableFeatureValuesConsidered}
\title{RSA model Kullback-Leibler divergence determination
(available feature values only)}
\usage{
RSAModelKLDiv3paramsOnlyAvailableFeatureValuesConsidered(
  data,
  par1,
  par2,
  par3
)
}
\arguments{
\item{data}{A matrix with data rows.

column structure:

[1:OC1,OC2,OC3,4:UUFeat, 5:Q1Feat,6:Q2Feat]

[7:Q1AnswerV1,V2,V3, 10:Q2AnswerV1,V2,V3]

\strong{1:OC1} Object 1. A value between 1 and 27.

\strong{2:OC2} Object 2. A value between 1 and 27.

\strong{3:OC3} Object 3. A value between 1 and 27.

\strong{4:UUFeat} Uttered feature. A number between 1 and 3. (1: shape, 2: pattern, 3: color)

\strong{5:Q1Feat} Questioned feature 1. A number between 1 and 3. (1: shape, 2: pattern, 3: color).

Example: If you utter "blue" (feature: color), then you can learn something about shape and texture preferences.

\strong{6:Q2Feat} Questioned feature 2. A number between 1 and 3. (1: shape, 2: pattern, 3: color).

Example: If you utter "blue" (feature: color), then you can learn something about shape and texture preferences.

\strong{7:Q1AnswerV1, V2, V3} The columns 7-9 contain the participants' slider values for the first questioned feature.

\strong{10:Q2AnswerV1, V2, C3} The columns 10-12 contain the participants' slider values for the second questioned feature.}

\item{par1}{\describe{
\item{softness parameter}{Specifies how much actual feature priorities come into play in the object choice (The larger the higher the tendency towards uniform liking).}
}}

\item{par2}{\describe{
\item{non-obedience parameter}{The extent to which the instruction of the speaker is obeyed by the listener.
(0 = full obedience, infinity = full instruction ignorance).}
}}

\item{par3}{\describe{
\item{alpha parameter}{An exponential scaling of the speaker choosing the utterance that maximizes the chance of the listener getting the target object right.}
}}
}
\value{
Minimized Kullback-Leibler divergence and the optimal parameters.
}
\description{
Full-RSA

The function calculates the optimal parameter values of the free parameters by estimating the log-likelihood of the
RSA model given model parameters and data. it also determines the actual RSA model Kullback-Leibler divergence.

3 parameter optimization considering only the available feature values present in the scene, i.e. feature values of shape, texture and color.
}
\details{
This function is used in the functions \code{\link{RSAModelLL1_1}}, \code{\link{RSAModelLL1_1_notObey.1}}, \code{\link{RSAModelLL1_2}}, \code{\link{RSAModelLL1_3}}, \code{\link{RSAModelLL2_n1}}, \code{\link{RSAModelLL2_n2}}, \code{\link{RSAModelLL2_n3}}, \code{\link{RSAModelLL3}}.
}
