% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSA_StratUttOptimization.R
\name{RSAModelLL1_2}
\alias{RSAModelLL1_2}
\title{Cost function for one parameter optimization.
Optimizing non-obedience.
Softness fixed at 0.}
\usage{
RSAModelLL1_2(params, data)
}
\arguments{
\item{params}{One value vector specifying one of the three parameters to be optimized:
\enumerate{
\item{softPrefValue parameter is fixed at 0, i.e. The strength of "preferring one entity over others". (The larger the value the higher the tendency towards uniform liking)}
\item{non-obedience parameter is optimized, i.e. The extent to which the instruction of the speaker is obeyed by the listener.
(0 = full obedience, infinity = full instruction ignorance).}
\item{alpha parameter is fixed at 1, i.e. An exponential scaling of the speaker choosing the utterance that maximizes the chance of the listener getting the target object right.}
}}

\item{data}{A matrix with data rows.

column structure:

[1:OC1,OC2,OC3,4:UUFeat, 5:Q1Feat,6:Q2Feat]

[7:Q1AnswerV1,V2,V3, 10:Q2AnswerV1,V2,V3]

\strong{1:OC1} Object 1. A value between 1 and 27.

\strong{2:OC2} Object 2. A value between 1 and 27.

\strong{3:OC3} Object 3. A value between 1 and 27.

\strong{4:UUFeat} Uttered feature. A number between 1 and 3. (1: shape, 2: pattern, 3: color)

\strong{5:Q1Feat} Questioned feature 1. A number between 1 and 3. (1: shape, 2: pattern, 3: color).

Example: If you utter "blue" (feature: color), then you can learn something about shape and texture preferences.

\strong{6:Q2Feat} Questioned feature 2. A number between 1 and 3. (1: shape, 2: pattern, 3: color).

Example: If you utter "blue" (feature: color), then you can learn something about shape and texture preferences.

\strong{7:Q1AnswerV1, V2, V3} The columns 7-9 contain the participants' slider values for the first questioned feature.

\strong{10:Q2AnswerV1, V2, C3} The columns 10-12 contain the participants' slider values for the second questioned feature.}
}
\value{
Minimized Kullback-Leibler divergence and the optimal parameters.
}
\description{
Full-RSA

1 parameter optimization; The non-obedience parameter is optimized.

The softness parameter and the alpha parameter are set to certain values.
}
\details{
This function uses \code{\link{RSAModelKLDiv3paramsOnlyAvailableFeatureValuesConsidered}}.
}
