% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSA_StratUttOptimization.R
\name{RSAModelLL2_n2}
\alias{RSAModelLL2_n2}
\title{Cost function for two parameter optimization.
Optimizing softness and alpha.
Non-obedience fixed at 0.}
\usage{
RSAModelLL2_n2(params, data)
}
\arguments{
\item{params}{Two value vector, which specifies two of three parameters (n1 = not the first, n2 = not the second, n3 = not the third) to be optimized:
\enumerate{
  \item{softPrefValue parameter is optimized, i.e. The strength of "preferring one entity over others". (The larger the value the higher the tendency towards uniform liking)}
  \item{non-obedience parameter is fixed at 0, i.e. The extent to which the instruction of the speaker is obeyed by the listener.
(0 = full obedience, infinity = full instruction ignorance).}
  \item{alpha parameter is optimized, i.e. An exponential scaling of the speaker choosing the utterance that maximizes the chance of the listener getting the target object right.}
}}

\item{data}{A matrix with data rows.

column structure:

[1:OC1,OC2,OC3,4:UUFeat, 5:Q1Feat,6:Q2Feat]

[7:Q1AnswerV1,V2,V3, 10:Q2AnswerV1,V2,V3]

\strong{1:OC1} Object 1. A value between 1 and 27.

\strong{2:OC2} Object 2. A value between 1 and 27.

\strong{3:OC3} Object 3. A value between 1 and 27.

\strong{4:UUFeat} Uttered feature. A number between 1 and 3. (1: shape, 2: pattern, 3: color)

\strong{5:Q1Feat} Questioned feature 1. A number between 1 and 3. (1: shape, 2: pattern, 3: color).

Example: If you utter "blue" (feature: color), then you can learn something about shape and texture preferences.

\strong{6:Q2Feat} Questioned feature 2. A number between 1 and 3. (1: shape, 2: pattern, 3: color).

Example: If you utter "blue" (feature: color), then you can learn something about shape and texture preferences.

\strong{7:Q1AnswerV1, V2, V3} The columns 7-9 contain the participants' slider values for the first questioned feature.

\strong{10:Q2AnswerV1, V2, C3} The columns 10-12 contain the participants' slider values for the second questioned feature.}
}
\value{
Minimized Kullback-Leibler divergence and the optimal parameters.
}
\description{
Full-RSA

2 parameter optimization; The softness and alpha parameters are optimized.

The non-obedience parameter is fixed.
}
\details{
This function uses \code{\link{RSAModelKLDiv3paramsOnlyAvailableFeatureValuesConsidered}}.
}
