% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSA_UttChoiceOptimization.R
\name{RSAModelUttKLDiv_4params}
\alias{RSAModelUttKLDiv_4params}
\title{Main cost function for four parameter optimization for the utterance choice experiments.
Optimizing softness, non-obedience, alpha and klValueFactor.}
\usage{
RSAModelUttKLDiv_4params(data, par1, par2, par3, par4)
}
\arguments{
\item{data}{A matrix with data rows.

column structure: \code{[1:OC1,OC2,OC3,4:numUttOptions,7-X:TurkerSliderValues]}

\strong{1:OC1} Object 1. A value between 1 and 27.

\strong{2:OC2} Object 2. A value between 1 and 27.

\strong{3:OC3} Object 3. A value between 1 and 27.

\strong{4:numUttOptions} The number of valid utterances in the scene.

\strong{7-X:TurkerSliderValues} These columns contain the participants' slider values.}

\item{par1}{\describe{
\item{softness parameter}{A parameter value between \code{[0,infinity)} (The larger the value the higher the tendency towards uniform liking).

Value reflects how categorical the listener's preferences are:

\strong{0:} The listener always picks her preferred object.

If the listener prefers \emph{red} objects, she will always pick the \emph{red} object in the scene.

\strong{infinity:} It is as likely for the listener to pick \emph{green}, \emph{blue} or \emph{red} objects.}}}

\item{par2}{\describe{
\item{non-obedience parameter}{Determines the extent to which the instruction of the speaker is obeyed by the listener.

(0 = full obedience, infinity = full instruction ignorance).

\strong{Example:}

\strong{0:} Listener always picks \emph{red} objects following the utterance \emph{"red"}.

\strong{infinity:} Listener as likely to pick \emph{green, blue} or \emph{red} objects even if the utterance is \emph{"red"}.}}}

\item{par3}{\describe{
\item{alpha parameter}{A parameter value between 0 and 1.

This parameter is an exponential scaling of the speaker choosing the
utterance that maximizes the chance of the listener getting the target object right.}
}}

\item{par4}{klValueFactor can be negative, 0 or positive:
\describe{
\item{\strong{zero}}{Don't care about learning about the feature preferences of the listener}
\item{\strong{positive}}{Care about learning about the feature preferences of the listener}
\item{\strong{negative}}{Trying to pick non-ambiguous utterances}
}}
}
\value{
Minimized Kullback-Leibler divergence and the optimal parameters.
}
\description{
Full-RSA

The actual RSA model Kullback-Leibler divergence determination for utterance choice experiments.

Softness, non-obedience, alpha and klValueFactor are optimized.
}
\details{
This function uses \code{\link{getBestInfGainUttPreferences}}.

And is used by

\code{\link{RSAModelUttKLDivParamD4}},

\code{\link{RSAModelUttKLDivParamD4A.2}},

\code{\link{RSAModelUttKLDivParamD4B.2}},

\code{\link{RSAModelUttKLDivParamD4AB.2}},

\code{\link{RSAModelUttKLDivParamAD4}},

\code{\link{RSAModelUttKLDivParamBD4}},

\code{\link{RSAModelUttKLDivParamAD4B.2}},

\code{\link{RSAModelUttKLDivParamBD4A.2}},

\code{\link{RSAModelUttKLDivParamCD4}},

\code{\link{RSAModelUttKLDivParamBCD4}},

\code{\link{RSAModelUttKLDivParamACD4}},

\code{\link{RSAModelUttKLDivParamABD4}},

\code{\link{RSAModelUttKLDivParamABCD4}}.
}
