% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSA_UttChoiceOptimization.R
\name{RSAModelUttKLDivParamAC}
\alias{RSAModelUttKLDivParamAC}
\title{Cost function for two parameter optimization for the utterance choice experiments.
Optimizing softness and klValueFactor.
Non-obedience is fixed.}
\usage{
RSAModelUttKLDivParamAC(params, data)
}
\arguments{
\item{params}{Two value vector specifying the two parameters to be optimized.
\enumerate{
\item{softness: A parameter value between \code{[0,infinity)} (The larger the value the higher the tendency towards uniform liking).

Value reflects how categorical the listener's preferences are:

\strong{0:} The listener always picks her preferred object.

If the listener prefers \emph{red} objects, she will always pick the \emph{red} object in the scene.

\strong{infinity:} It is as likely for the listener to pick \emph{green}, \emph{blue} or \emph{red} objects.}
\item{klValueFactor: A parameter that can be negative, 0 or positive (Here it is set to = 1):
\describe{
\item{\strong{zero}}{Don't care about learning about the feature preferences of the listener}
\item{\strong{positive}}{Care about learning about the feature preferences of the listener}
\item{\strong{negative}}{Trying to pick non-ambiguous utterances}
}}
}}

\item{data}{A matrix with data rows.

column structure: \code{[1:OC1,OC2,OC3,4:numUttOptions,7-X:TurkerSliderValues]}

\strong{1:OC1} Object 1. A value between 1 and 27.

\strong{2:OC2} Object 2. A value between 1 and 27.

\strong{3:OC3} Object 3. A value between 1 and 27.

\strong{4:numUttOptions} The number of valid utterances in the scene.

\strong{7-X:TurkerSliderValues} These columns contain the participants' slider values.}
}
\value{
Minimized Kullback-Leibler divergence and the optimal parameters.
}
\description{
Full-RSA

Softness and klValueFactor are optimized.

Non-obedience is fixed at 0.
}
\details{
This function uses \code{\link{RSAModelUttKLDiv_3params}}.
}
