% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSA_UttChoiceOptimization.R
\name{RSAModelUttKLDivParamD4A.2}
\alias{RSAModelUttKLDivParamD4A.2}
\title{Cost function for one parameter optimization for the utterance choice experiments.
Optimizing klValueFactor.
Softness, non-obedience and alpha are fixed.}
\usage{
RSAModelUttKLDivParamD4A.2(params, data)
}
\arguments{
\item{params}{One value vector specifying the parameter to be optimized.
\enumerate{
\item{klValueFactor can be negative, zero or positive:
\describe{
\item{zero}{Don't care about learning about feature preferences of the listener}
\item{positive}{Care about learning about feature preferences of the listener}
\item{negative}{Trying to pick non-ambiguous utterances}
}}
}}

\item{data}{A matrix with data rows.

column structure: \code{[1:OC1,OC2,OC3,4:numUttOptions,7-X:TurkerSliderValues]}

\strong{1:OC1} Object 1. A value between 1 and 27.

\strong{2:OC2} Object 2. A value between 1 and 27.

\strong{3:OC3} Object 3. A value between 1 and 27.

\strong{4:numUttOptions} The number of valid utterances in the scene.

\strong{7-X:TurkerSliderValues} These columns contain the participants' slider values.}
}
\value{
Minimized Kullback-Leibler divergence and the optimal parameters.
}
\description{
Full-RSA

Softness is fixed at 0.2.

non-obedience is fixed at 0.

Alpha is  fixed at to 1.
}
\details{
This function uses \code{\link{RSAModelUttKLDiv_4params}}.
}
