% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRSA_StratUtt.R
\name{simplePragmaticSpeaker}
\alias{simplePragmaticSpeaker}
\title{Simple pragmatic speaker function}
\usage{
simplePragmaticSpeaker(
  utterance,
  obj,
  preferencesPrior,
  validUtterances,
  currentObjects,
  uttToObjProbs,
  objectPreferenceSoftPriors
)
}
\arguments{
\item{utterance}{The uttered word by the speaker that the listener hears.

An index referring to one of the values in the vector validUtterances.}

\item{obj}{The object chosen by the listener. A value referring to the index 1,2 or 3.}

\item{preferencesPrior}{A vector of length 9.

Probability mass over all feature values present in the scenario plus a "no preference" case.

Gives a prior preferences distribution over all (nine) feature values.}

\item{validUtterances}{A vector of utterances that correspond to all feature values present
in the current objects in the scene.

For example, it only makes sense to utter \emph{"red"} in a scene if there are \emph{red} objects present.}

\item{currentObjects}{A vector of three values in \code{{1,...,27}} specifying the target and the other two objects in the scene.

The target is the first object in the vector \code{(index = 1)}.}

\item{uttToObjProbs}{A matrix. The rows map each possible utterance that corresponds to each present feature value
of the current objects. The columns represent the three objects in the scene.

This reflects the obedience-parameter and which objects match the respective utterance.
The matrix shows the probability that a certain object is chosen following a certain utterance, that is valid in the scene.
The number of rows of the matrix match the length of the validUtterances vector.}

\item{objectPreferenceSoftPriors}{A list of preference priors for all valid utterances based on the object in the scene.

The list has as many rows as the length of the validUtterances vector + 1.

Each row in the list contains a vector of length 3, as there are three objects in the scene.

The extra row is for the case of no feature preferences whatsoever, i.e. uniform prior over all three objects in the scene.}
}
\value{
A vector with the same as the validUtterances vector.

Normalized posterior probability over preferences- given the utterance, the object choice by the listener, and prior over preferences of the listener.
}
\description{
Simple RSA

The simple pragmatic speaker considers all "imaginable" (i.e. implemented)
preference distributions over objects of the listener.
It starts with a prior assumption over the possible listener's preferences.
It then infers the posterior over these preferences given the listener makes a particular object choice.
P(listener's feature value preferences | utterance, object choice by the listener,
 prior over listener's feature value preferences)
}
\examples{
\donttest{simplePragmaticSpeaker(utterance, obj, preferencesPrior,
validUtterances, currentObjects, uttToObjProbs, objectPreferenceSoftPriors)}

output:
[1] 0.17 0.17 0.17 0.17 0.17 0.17

}
