% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRSA_UttChoiceOptimization_iterative.R
\name{SimpleRSAModelUttKLDiv_3params_independent}
\alias{SimpleRSAModelUttKLDiv_3params_independent}
\title{Main cost function for three parameter optimization for the utterance choice experiments.
Optimizing softness, non-obedience and klValueFactor. (independent and iterative setting)}
\usage{
SimpleRSAModelUttKLDiv_3params_independent(data, par1, par2, par3)
}
\arguments{
\item{data}{A matrix with data rows.

column structure: \code{[1:OC1,OC2,OC3,4:numUttOptions,6: picked utterance]}

\strong{1:OC1} Object 1. A value between 1 and 27.

\strong{2:OC2} Object 2. A value between 1 and 27.

\strong{3:OC3} Object 3. A value between 1 and 27.

\strong{4:numUttOptions} The number of valid utterances in the scene.

\strong{6: picked utterance} The utterance picked by the participant.}

\item{par1}{\describe{
\item{softness parameter}{Specifies how much actual feature priorities come into play in the object choice (The larger the higher the tendency towards uniform liking).}
}}

\item{par2}{\describe{
\item{non-obedience parameter}{The extent to which the instruction of the speaker is obeyed by the listener.
(0 = full obedience, infinity = full instruction ignorance).}
}}

\item{par3}{\describe{
\item{klValueFactor}{A parameter that can be negative, 0 or positive:}
\describe{
\item{\strong{zero}}{Don't care about learning about the feature preferences of the listener}
\item{\strong{positive}}{Care about learning about the feature preferences of the listener}
\item{\strong{negative}}{Trying to pick non-ambiguous utterances}
}}}
}
\value{
Minimized Kullback-Leibler divergence and the optimal parameters.
}
\description{
Simple RSA

Actual RSA model Kullback-Leibler divergence determination for utterance choice experiments.

Softness, non-obedience and klValueFactor are optimized.
}
\details{
This function uses \code{\link{getSimpleBestInfGainUttPreferencesIterative}}.

This function is used in

\code{\link{SimpleRSAModelUttKLDivParamA_independent}},

\code{\link{SimpleRSAModelUttKLDivParamB_independent}},

\code{\link{SimpleRSAModelUttKLDivParamK_independent}},

\code{\link{SimpleRSAModelUttKLDivParamBK_independent}},

\code{\link{SimpleRSAModelUttKLDivParamAK_independent}},

\code{\link{SimpleRSAModelUttKLDivParamAK.2_independent}},

\code{\link{SimpleRSAModelUttKLDivParamA.2_independent}},

\code{\link{SimpleRSAModelUttKLDivParamB.2_independent}},

\code{\link{SimpleRSAModelUttKLDivParamK.2.2_independent}},

\code{\link{SimpleRSAModelUttKLDivParamK.2.0_independent}},

\code{\link{SimpleRSAModelUttKLDivParamBK.2_independent}},

\code{\link{SimpleRSAModelUttKLDivParamAB_independent}},

\code{\link{SimpleRSAModelUttKLDivParamABK_independent}}.
}
