% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRSA_UttChoiceOptimization_iterative.R
\name{SimpleRSAModelUttKLDivParamAK_iterative}
\alias{SimpleRSAModelUttKLDivParamAK_iterative}
\title{Cost function for two parameter optimization for the utterance choice experiments. (iterative setting)
Optimizing softness and klValueFactor.
Non-obedience is fixed.}
\usage{
SimpleRSAModelUttKLDivParamAK_iterative(params, data)
}
\arguments{
\item{params}{Two value vector specifying the two parameters to be optimized.
\enumerate{
\item{softPrefValue, i.e., the strength of preferring one entity over others.}
\item{klValueFactor can be negative, zero or positive:
\describe{
\item{zero}{Don't care about learning about feature preferences of the listener}
\item{positive}{Care about learning about feature preferences of the listener}
\item{negative}{Trying to pick non-ambiguous utterances}
}}}}

\item{data}{A matrix with data rows.

column structure: \code{[1:OC1,OC2,OC3,4:numUttOptions,6: picked utterance]}

\strong{1:OC1} Object 1. A value between 1 and 27.

\strong{2:OC2} Object 2. A value between 1 and 27.

\strong{3:OC3} Object 3. A value between 1 and 27.

\strong{4:numUttOptions} The number of valid utterances in the scene.

\strong{6: picked utterance} The utterance picked by the participant.}
}
\value{
Minimized Kullback-Leibler divergence and the optimal parameters.
}
\description{
Simple RSA

Softness and klValueFactor are optimized.

Non-obedience is fixed at 0.
}
\details{
This is the iterative version of \code{\link{SimpleRSAModelUttKLDivParamAK}}

This function uses \code{\link{SimpleRSAModelUttKLDiv_3params_iterative}}.
}
