% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRSA_UttChoiceOptimization_iterative.R
\name{SimpleRSAModelUttKLDivParamBK.2_iterative}
\alias{SimpleRSAModelUttKLDivParamBK.2_iterative}
\title{Cost function for two parameter optimization for the utterance choice experiments. (iterative setting)
Optimizing non-obedience and klValueFactor.
Softness is fixed.}
\usage{
SimpleRSAModelUttKLDivParamBK.2_iterative(params, data)
}
\arguments{
\item{params}{Two value vector specifying two of the three parameters to be optimized.
\enumerate{
\item{non-obedience: Determines the extent to which the instruction of the speaker is obeyed by the listener.

(0 = full obedience, infinity = full instruction ignorance).

\strong{Example:}

\strong{0:} Listener always picks \emph{red} objects following the utterance \emph{"red"}.

\strong{infinity:} Listener as likely to pick \emph{green, blue} or \emph{red} objects even if the utterance is \emph{"red"}.}
\item{klValueFactor: A parameter that can be negative, 0 or positive:
\describe{
\item{zero}{Don't care about learning about feature preferences of the listener}
\item{positive}{Care about learning about feature preferences of the listener}
\item{negative}{Trying to pick non-ambiguous utterances}
}}}}

\item{data}{A matrix with data rows.

column structure: \code{[1:OC1,OC2,OC3,4:numUttOptions,6: picked utterance]}

\strong{1:OC1} Object 1. A value between 1 and 27.

\strong{2:OC2} Object 2. A value between 1 and 27.

\strong{3:OC3} Object 3. A value between 1 and 27.

\strong{4:numUttOptions} The number of valid utterances in the scene.

\strong{6: picked utterance} The utterance picked by the participant.}
}
\value{
Minimized Kullback-Leibler divergence and the optimal parameters.
}
\description{
Simple RSA

Non-obedience and klValueFactor are optimized.

Softness is fixed at 0.2.
}
\details{
This is the iterative version of \code{\link{SimpleRSAModelUttKLDivParamBK.2}}.

This function uses \code{\link{SimpleRSAModelUttKLDiv_3params_iterative}}.
}
