% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawData.R
\docType{data}
\name{X3_rawData}
\alias{X3_rawData}
\title{Experiment 2 raw data (Epistemic utterance choice)}
\format{
\describe{
  \item{workerid}{Participant's ID}
  
  \item{item}{Unique ID of a stimulus}
  
  \item{slide_number}{Indicates trial order.}
  
  \item{condition}{A six-digit code that infers to which category the item belongs. 
  In this experiment, it is only necessary to identify classes that have 
  potentially ambiguous utterances.}
  
  \item{language}{Participant self-reported native language}
  
  \item{pref1-pref9}{Utterance 1-9.}
  
  \item{response1- response9}{The value to which the slider (1-9) was adjusted 
  by the participant. Indicates how useful an utterance is for finding out 
  listener preferences. The corresponding property that they are rating is 
  stored under pref1-9. Can be in range 0−1.
  
  Small numbers mean less useful, larger numbers mean more useful.}
  
  \item{obj1, obj2, obj3}{A three-digit code referring to the object that the 
  speaker intends to signal by the utterance (obj1), or to one of the other 
  two objects present in the scene (obj2, obj3).
  
  The first digit of the code refers to shape, the second digit to texture, 
  and the third digit to color of the object.
  
  \strong{1 Shape:} 1 cloud, 2 circle, 3 square;
  
  \strong{2 Texture:} 1 solid, 2 striped, 3 polka-dotted;
  
  \strong{3 Color:} 1 blue, 2 red, 3 green.
  
  \strong{Example:} If the target is a solid blue square the code would be: 311.}
  
  \item{ambiguous}{Marks whether a trial contains any ambiguous utterances.}
  
  \item{numFeatures}{The number of feature values present in a scene. 
  
  For example, if we have 3 red objects, all of them solid, 1 square, 1 cloud, 
  1 circle, then the number of features would be 5.
  
  Value in range 3-9.}
}
}
\usage{
X3_rawData
}
\description{
This file contains the raw data from the utterance choice experiment.
}
\keyword{datasets}
